+++ src/onegov.core/HISTORY.rst
@@ -3,6 +3,10 @@ Changelog
 
 Unreleased
 ~~~~~~~~~~
+
+- Adds a distributed lock mechanism using postgres.
+  [href]
+
 0.12.3 (2015-12-21)
 ~~~~~~~~~~~~~~~~~~~
 
@@ -105,7 +109,7 @@ Unreleased
 ~~~~~~~~~~~~~~~~~~~
 
 - Adjacency lists are now always ordered by the value in their 'order' column.
-  
+
   When adding new items to a parent, A-Z is enforced between the children, as
   long as the children are already sorted A-Z. Once this holds no longer true,
   no sorting will be imposed on the unsorted children until they are sorted